requirements.txt
# Django & Django-rest-framework
Django>=4.2,<5
djangorestframework==3.15.1
django-cors-headers==4.3.0

# Token Authentication
djangorestframework-simplejwt==5.2.2

# Database
psycopg2-binary==2.9.6

# API Documentation
drf-yasg==1.21.7

# Static Files
whitenoise==6.6.0

# Config
python-decouple==3.8

# Shell
ipython==8.14.0

# Load balancing
gunicorn==22.0.0
